collector_service/requirements.txt
Flask==2.1.3
Werkzeug==2.0.3
paho-mqtt==1.6.1
kafka-python==2.0.3
